improve ff tests
